feat(safety_harness/transaction): implement atomic multi-file edit transactions with verification
- Introduce two-phase commit for edit transactions: prepare writes to temp files; commit performs atomic renames.
- Track file existence to support rollback that restores original state, including removing newly created files.
- Create configurable syntactic and semantic verification locks for test scenarios.
- Moves text edit application logic to verification/apply.rs with robust handling of LF and CRLF line endings.
- Add extensive tests covering success, syntactic and semantic lock failures, backend errors, new file creation, and multi-file edits.
- Update docs to reflect verification and atomic transaction behavior and acceptance criteria.

This change enforces safety harness guarantees that multi-file edits are either fully committed or fully rolled back on failure, with verification locks gating commits.
